+++ data/src/main/java/com/yapp/growth/data/repository/ConfirmPlanRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.yapp.growth.data.repository
 
 import com.yapp.growth.data.source.ConfirmPlanDataSource
 import com.yapp.growth.domain.NetworkResult
-import com.yapp.growth.domain.entity.RespondUsers
+import com.yapp.growth.domain.entity.ResponsePlan
 import com.yapp.growth.domain.repository.ConfirmPlanRepository
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -12,7 +12,7 @@ internal class ConfirmPlanRepositoryImpl @Inject constructor(
     private val dataSource: ConfirmPlanDataSource
 ) : ConfirmPlanRepository {
 
-    override suspend fun getRespondUsers(promisingKey: Long): NetworkResult<RespondUsers> {
+    override suspend fun getRespondUsers(promisingKey: Long): NetworkResult<ResponsePlan> {
         return dataSource.getRespondUsers(promisingKey)
     }
