+++ data/src/main/java/com/yapp/growth/data/source/ConfirmPlanDataSource.kt
@@ -1,9 +1,9 @@
 package com.yapp.growth.data.source
 
 import com.yapp.growth.domain.NetworkResult
-import com.yapp.growth.domain.entity.RespondUsers
+import com.yapp.growth.domain.entity.ResponsePlan
 
 interface ConfirmPlanDataSource {
 
-    suspend fun getRespondUsers(promisingKey: Long): NetworkResult<RespondUsers>
+    suspend fun getRespondUsers(promisingKey: Long): NetworkResult<ResponsePlan>
 }
