+++ domain/src/main/java/com/yapp/growth/domain/entity/SendingResponsePlan.kt
@@ -1,8 +1,6 @@
 package com.yapp.growth.domain.entity
 
-import java.util.*
-
-data class RespondPlan(
+data class SendingResponsePlan(
     val date: String,
     val hours: List<String>,
     val timeList: MutableList<Boolean>
