+++ domain/src/main/java/com/yapp/growth/domain/repository/ConfirmPlanRepository.kt
@@ -1,9 +1,9 @@
 package com.yapp.growth.domain.repository
 
 import com.yapp.growth.domain.NetworkResult
-import com.yapp.growth.domain.entity.RespondUsers
+import com.yapp.growth.domain.entity.ResponsePlan
 
 interface ConfirmPlanRepository {
 
-    suspend fun getRespondUsers(promisingKey: Long): NetworkResult<RespondUsers>
+    suspend fun getRespondUsers(promisingKey: Long): NetworkResult<ResponsePlan>
 }
