+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/manage/confirm/ConfirmPlanViewModel.kt
@@ -3,8 +3,8 @@ package com.yapp.growth.presentation.ui.main.manage.confirm
 import androidx.lifecycle.viewModelScope
 import com.yapp.growth.base.BaseViewModel
 import com.yapp.growth.domain.NetworkResult
-import com.yapp.growth.domain.entity.RespondPlan
-import com.yapp.growth.domain.entity.RespondUsers
+import com.yapp.growth.domain.entity.SendingResponsePlan
+import com.yapp.growth.domain.entity.ResponsePlan
 import com.yapp.growth.domain.entity.User
 import com.yapp.growth.domain.usecase.GetRespondUsersUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -23,16 +23,16 @@ class ConfirmPlanViewModel @Inject constructor(
     ConfirmPlanViewState()
 ) {
 
-    private val _dates = MutableStateFlow<List<RespondPlan>>(emptyList())
-    val dates: StateFlow<List<RespondPlan>>
-        get() = _dates.asStateFlow()
+    private val _sendResponsePlan = MutableStateFlow<List<SendingResponsePlan>>(emptyList())
+    val sendResponsePlan: StateFlow<List<SendingResponsePlan>>
+        get() = _sendResponsePlan.asStateFlow()
 
-    private val _respondUser = MutableStateFlow<RespondUsers>(RespondUsers(emptyList(),
+    private val _responsePlan = MutableStateFlow<ResponsePlan>(ResponsePlan(emptyList(),
         emptyList(), emptyList(),0,"","", emptyList(), emptyList()
     ))
 
-    val respondUser: StateFlow<RespondUsers>
-        get() = _respondUser
+    val responsePlan: StateFlow<ResponsePlan>
+        get() = _responsePlan
 
     private val _currentClickTimeIndex = MutableStateFlow(-1 to -1)
     val currentClickTimeIndex: StateFlow<Pair<Int, Int>>
@@ -51,33 +51,33 @@ class ConfirmPlanViewModel @Inject constructor(
             val result = (getRespondUsersUseCase(promisingKey) as? NetworkResult.Success)?.data
             result?.let {
                 makeRespondList(it)
-                _respondUser.value = it
+                _responsePlan.value = it
             }
         }
     }
 
-    private fun makeRespondList(data: RespondUsers) {
+    private fun makeRespondList(data: ResponsePlan) {
         val booleanArray = Array(data.totalCount*2) { false }
 
-        val temp = mutableListOf<RespondPlan>().also { list ->
-            repeat(data.avaliableDate.size) {
-                list.add(RespondPlan(
-                    date = data.avaliableDate[it],
+        val temp = mutableListOf<SendingResponsePlan>().also { list ->
+            repeat(data.availableDate.size) {
+                list.add(SendingResponsePlan(
+                    date = data.availableDate[it],
                     hours = data.hourList,
                     timeList = booleanArray.copyOf().toMutableList()
                 ))
             }
         }.toList()
 
-        _dates.value = temp
+        _sendResponsePlan.value = temp
     }
 
     private fun filterCurrentSelectedUser(dateIndex: Int, minuteIndex: Int) {
         viewModelScope.launch(Dispatchers.Default) {
-            val day = respondUser.value.avaliableDate[dateIndex]
-            var hour = respondUser.value.hourList[minuteIndex/2]
+            val day = responsePlan.value.availableDate[dateIndex]
+            var hour = responsePlan.value.hourList[minuteIndex/2]
 
-            val blockList = respondUser.value.timeTable.find { it.date == day }?.blocks
+            val blockList = responsePlan.value.timeTable.find { it.date == day }?.blocks
             val userList = blockList?.let { block ->
                 block.find { it.index == minuteIndex }?.users
             }
@@ -95,7 +95,7 @@ class ConfirmPlanViewModel @Inject constructor(
             is ConfirmPlanEvent.OnClickPreviousDayButton -> { }
             is ConfirmPlanEvent.OnClickConfirmButton -> { }
             is ConfirmPlanEvent.OnClickTimeTable -> {
-                _dates.value[event.dateIndex].timeList[event.minuteIndex] = _dates.value[event.dateIndex].timeList[event.minuteIndex].not()
+                _sendResponsePlan.value[event.dateIndex].timeList[event.minuteIndex] = _sendResponsePlan.value[event.dateIndex].timeList[event.minuteIndex].not()
                 _currentClickTimeIndex.value  = event.dateIndex to event.minuteIndex
                 filterCurrentSelectedUser(event.dateIndex, event.minuteIndex)
             }
