+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/manage/respond/RespondPlanViewModel.kt
@@ -3,8 +3,8 @@ package com.yapp.growth.presentation.ui.main.manage.respond
 import androidx.lifecycle.viewModelScope
 import com.yapp.growth.base.BaseViewModel
 import com.yapp.growth.domain.NetworkResult
-import com.yapp.growth.domain.entity.RespondPlan
-import com.yapp.growth.domain.entity.RespondUsers
+import com.yapp.growth.domain.entity.SendingResponsePlan
+import com.yapp.growth.domain.entity.ResponsePlan
 import com.yapp.growth.domain.usecase.GetRespondUsersUseCase
 import com.yapp.growth.presentation.ui.main.manage.respond.RespondPlanContract.*
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -12,9 +12,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import java.text.DateFormat
-import java.text.SimpleDateFormat
-import java.util.*
 import javax.inject.Inject
 
 @HiltViewModel
@@ -23,18 +20,18 @@ class RespondPlanViewModel @Inject constructor(
 ): BaseViewModel<RespondPlanViewState, RespondPlanSideEffect, RespondPlanEvent>(
     RespondPlanViewState()
 ) {
-    private val _dates = MutableStateFlow<List<RespondPlan>>(emptyList())
-    val dates: StateFlow<List<RespondPlan>>
-        get() = _dates.asStateFlow()
+    private val _sendResponsePlan = MutableStateFlow<List<SendingResponsePlan>>(emptyList())
+    val sendResponsePlan: StateFlow<List<SendingResponsePlan>>
+        get() = _sendResponsePlan.asStateFlow()
 
-    private val _respondUser = MutableStateFlow<RespondUsers>(
-        RespondUsers(
+    private val _responsePlan = MutableStateFlow<ResponsePlan>(
+        ResponsePlan(
             emptyList(), emptyList(), emptyList(), 0, "", "", emptyList(), emptyList()
         )
     )
 
-    val respondUser: StateFlow<RespondUsers>
-        get() = _respondUser
+    val responsePlan: StateFlow<ResponsePlan>
+        get() = _responsePlan
 
     private val _clickCount = MutableStateFlow(0)
     val clickCount: StateFlow<Int>
@@ -49,32 +46,32 @@ class RespondPlanViewModel @Inject constructor(
             val result = (getRespondUsersUseCase(promisingKey) as? NetworkResult.Success)?.data
             result?.let {
                 makeRespondList(it)
-                _respondUser.value = it
+                _responsePlan.value = it
             }
         }
     }
 
-    private fun makeRespondList(data: RespondUsers) {
+    private fun makeRespondList(data: ResponsePlan) {
         val booleanArray = Array(data.totalCount*2) { false }
 
-        val temp = mutableListOf<RespondPlan>().also { list ->
-            repeat(data.avaliableDate.size) {
-                list.add(RespondPlan(
-                    date = data.avaliableDate[it],
+        val temp = mutableListOf<SendingResponsePlan>().also { list ->
+            repeat(data.availableDate.size) {
+                list.add(SendingResponsePlan(
+                    date = data.availableDate[it],
                     hours = data.hourList,
                     timeList = booleanArray.copyOf().toMutableList()
                 ))
             }
         }.toList()
 
-        _dates.value = temp
+        _sendResponsePlan.value = temp
     }
 
     override fun handleEvents(event: RespondPlanEvent) {
         when (event) {
             is RespondPlanEvent.OnClickTimeTable -> {
-                _dates.value[event.dateIndex].timeList[event.minuteIndex] = _dates.value[event.dateIndex].timeList[event.minuteIndex].not()
-                if (dates.value[event.dateIndex].timeList[event.minuteIndex]) _clickCount.value += 1 else _clickCount.value -= 1
+                _sendResponsePlan.value[event.dateIndex].timeList[event.minuteIndex] = _sendResponsePlan.value[event.dateIndex].timeList[event.minuteIndex].not()
+                if (sendResponsePlan.value[event.dateIndex].timeList[event.minuteIndex]) _clickCount.value += 1 else _clickCount.value -= 1
             }
             is RespondPlanEvent.OnClickNextDayButton -> {  }
             is RespondPlanEvent.OnClickPreviousDayButton -> {  }
